Apply patch from ticket:165 -- changelog:
Attached is a patch that allows hachoir-http to parse arbitrary streams. This
is achieved by using a "dependency"-like system on streams: streams will depend
on the stream which created it (by including that stream in their stream data).
This is more efficient space-wise than the storage of the tags and data (which
was the old way) but will take a bit longer to load, as hachoir-http must parse
up to the previous point in the stream, parse the substream, and possibly do
this a number of times before arriving at the requested stream.


git-svn-id: http://hachoir.org/svn/trunk@2647 2c54552e-060f-0410-81f1-f74aad87cf62